requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
enum-compat==0.0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Naked==0.1.31
pycryptodomex==3.9.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
